+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v0.4.0-dev (Apr 07, 2026)
 
+- Calculate basket after OTP verificatiion (@W-21918545)
 - Update accessibility baseline and fix accessibility violations on order list page (@W-21685164)
 - Consolidate `action.set-site`, `action.set-locale`, and `action.set-currency` into a single `action.set-site-context` route (@W-21787262) ([#1342](https://github.com/commerce-emu/storefront-next/pull/1342))
 - Clean up RSC/React Server Components references from documentation, comments, and config ([#1363](https://github.com/commerce-emu/storefront-next/pull/1363))